adicionando security na loja de games

+++ src/main/java/com/generation/lojagames/controller/CategoriaController.java
@@ -1,6 +1,7 @@
 package com.generation.lojagames.controller;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -13,69 +14,64 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 import com.generation.lojagames.model.Categoria;
 import com.generation.lojagames.repository.CategoriaRepository;
-import com.generation.lojagames.repository.ProdutoRepository;
 
 import jakarta.validation.Valid;
-
 
 @RestController
 @RequestMapping("/categorias")
 @CrossOrigin(origins = "*", allowedHeaders = "*")
 public class CategoriaController {
-	
-    @Autowired
-    private CategoriaRepository categoriaRepository;
-    
+	// Conectando com a Repository
 	@Autowired
-	private ProdutoRepository produtoRepository;
+	private CategoriaRepository categoriaRepository;
 	
-
-	@GetMapping
+	// Métodos CRUD básicos
+	@PostMapping("/cadastrar")
+	public ResponseEntity<Categoria> post(@Valid @RequestBody Categoria categoria){
+		return ResponseEntity.status(HttpStatus.CREATED).body(categoriaRepository.save(categoria));
+	}
+	
+	@GetMapping("/all")
 	public ResponseEntity<List<Categoria>> getAll(){
 		return ResponseEntity.ok(categoriaRepository.findAll());
 	}
 	
-	@GetMapping("/{id}")
-	public ResponseEntity<Categoria> getById(@PathVariable Long id) {
-	    return categoriaRepository.findById(id)
-	            .map(categoria -> ResponseEntity.ok(categoria))
-	            .orElse(ResponseEntity.status(HttpStatus.NOT_FOUND).build());
-	}
-	
-	@GetMapping("/tipo/{tipo}")
-	public ResponseEntity<List<Categoria>> getByTitulo(@PathVariable String tipo) {
-	    List<Categoria> categoria = categoriaRepository.findAllByTipoContainingIgnoreCase(tipo);
-	    /*IF ternario*/
-	    return categoria.isEmpty() ? ResponseEntity.status(HttpStatus.NOT_FOUND).build() : ResponseEntity.status(HttpStatus.OK).body(categoria);
-	}
-	
-	@PostMapping
-	public ResponseEntity<Categoria> post(@RequestBody Categoria categoria) {
-	    return ResponseEntity.status(HttpStatus.CREATED).body(categoriaRepository.save(categoria));
-	}
-	
-	@PutMapping
-	public ResponseEntity<?> put(@Valid @RequestBody Categoria categoria){
-	    return categoriaRepository.findById(categoria.getId())
-	            .map(resposta -> {
-	            	categoriaRepository.save(categoria);
-	                return ResponseEntity.status(HttpStatus.OK).build();
-	            })
-	            .orElse(ResponseEntity.status(HttpStatus.BAD_REQUEST).build());
-	}
-	
-	@DeleteMapping("/{id}")
-	public ResponseEntity<?> deleteById(@PathVariable Long id) { 
-		return categoriaRepository.findById(id)
-				.map(resposta -> {
-					categoriaRepository.deleteById(id);
-					return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
-				})
+	@PutMapping("/atualizar")
+	public ResponseEntity<Categoria> put(@Valid @RequestBody Categoria categoria){
+		return categoriaRepository.findById(categoria.getId())
+				.map(resposta -> ResponseEntity.status(HttpStatus.OK).body(categoriaRepository.save(categoria)))
 				.orElse(ResponseEntity.status(HttpStatus.NOT_FOUND).build());
-
 	}
+	
+	@ResponseStatus(HttpStatus.NO_CONTENT)
+	@DeleteMapping("/{id}")
+	public void delete(@PathVariable Long id) {
+		Optional<Categoria> optional = categoriaRepository.findById(id);
+		
+		if (optional.isEmpty()) {
+			throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+		}
+		else {
+			categoriaRepository.deleteById(id);
+		}
+	}
+	
+	// Métodos Adicionais
+		@GetMapping("/{id}")
+		public ResponseEntity<Categoria> getById(@PathVariable Long id){
+			return categoriaRepository.findById(id)
+					.map(resposta -> ResponseEntity.ok(resposta))
+					.orElse(ResponseEntity.status(HttpStatus.NOT_FOUND).build());
+		}
+		
+		@GetMapping("/nome/{nome}")
+		public ResponseEntity<List<Categoria>> getByNome(@PathVariable String nome){
+			return ResponseEntity.ok(categoriaRepository.findAllByNomeContainingIgnoreCase(nome));
+		}
 }
